requirements.txt
pandas==2.1.0
requests==2.31.0
requests-html==0.10.0
requests-toolbelt==1.0.0
numpy==1.25.2
bs4==0.0.1
apache-airflow==2.9.0
PyGithub
pymongo
overpy